+++ matzipback/src/main/java/com/matzip/matzipback/matzipList/command/domain/aggregate/Review.java
@@ -17,9 +17,9 @@ public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reviewSeq;
-    private int reviewUserSeq;
-    private int restaurantSeq;
+    private long reviewSeq;
+    private long reviewUserSeq;
+    private long restaurantSeq;
     private String reviewContent;
     private String reviewStatus;
     @CreatedDate
